refactor : answer삭제 시 유효성 검증

+++ src/main/java/qna/domain/Answer.java
@@ -1,5 +1,6 @@
 package qna.domain;
 
+import qna.CannotDeleteException;
 import qna.NotFoundException;
 import qna.UnAuthorizedException;
 
@@ -44,7 +45,9 @@ public class Answer extends AbstractEntity {
         this.contents = contents;
     }
 
-    public void delete(DeleteHistories deleteHistories) {
+    public void delete(User loginUser, DeleteHistories deleteHistories) throws CannotDeleteException {
+        checkAnswersOwnerIsLoginUser(loginUser);
+
         this.deleted = true;
         deleteHistories.addDeleteHistory(new DeleteHistory(ContentType.ANSWER, getId(), writer));
     }
@@ -72,5 +75,11 @@ public class Answer extends AbstractEntity {
     @Override
     public String toString() {
         return "Answer [id=" + getId() + ", writer=" + writer + ", contents=" + contents + "]";
+    }
+
+    public void checkAnswersOwnerIsLoginUser(User loginUser) throws CannotDeleteException {
+        if (!isOwner(loginUser)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
     }
 }
